+++ lib/modules/type_of_wash/types_of_wash_screen.dart
@@ -23,17 +23,24 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
         child: Padding(
           padding: EdgeInsets.all(MediaQuery.of(context).size.height * 0.05),
           child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
-                  Text("Types Of Washes",
-                      style: GoogleFonts.inter(
-                          color: const Color(0xff333333),
-                          fontSize: 20,
-                          fontWeight: FontWeight.w600)),
+                  SizedBox(
+                    child: Row(
+                      children: [
+                        BackButton(),
+                        SizedBox(width: MediaQuery.of(context).size.width*0.005),
+                        Text("Types Of Washes",
+                            style: GoogleFonts.inter(
+                                color: const Color(0xff333333),
+                                fontSize: 20,
+                                fontWeight: FontWeight.w600)),
+                      ],
+                    ),
+                  ),
                   ElevatedButton(
                     onPressed: () {
                       Navigator.push(
@@ -73,7 +80,7 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                           gridDelegate:
                               const SliverGridDelegateWithFixedCrossAxisCount(
                                   crossAxisCount: 3,
-                                  childAspectRatio: 1.8,
+                                  childAspectRatio: 200 / 105,
                                   mainAxisSpacing: 20.0, // spacing between rows
                                   crossAxisSpacing: 20.0),
                           itemBuilder: (context, index) {
@@ -93,8 +100,7 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                                 color: Color(int.parse(
                                     snapshot.data!.docs[index]['bgColor'])),
                                 child: Column(
-                                    mainAxisAlignment:
-                                        MainAxisAlignment.spaceEvenly,
+                                    crossAxisAlignment: CrossAxisAlignment.start,
                                     children: [
                                       Padding(
                                         padding: const EdgeInsets.all(8.0),
@@ -102,11 +108,35 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                                           alignment: Alignment.bottomRight,
                                           child: IconButton(
                                               onPressed: () {
-                                                FirebaseFirestore.instance
-                                                    .collection("typeOfWashes")
-                                                    .doc(snapshot
-                                                        .data!.docs[index].id)
-                                                    .delete();
+                                                showDialog(
+                                                    context: context,
+                                                    builder: (context) {
+                                                      return CupertinoAlertDialog(
+                                                        title: const Text("Delete"),
+                                                        content: const Text("Are you sure ypu want to delete this?"),
+                                                        actions: <CupertinoDialogAction>[
+                                                          CupertinoDialogAction(
+                                                            onPressed: () {
+                                                              Navigator.pop(context);
+                                                            },
+                                                            child: const Text("No"),
+                                                          ),
+                                                          CupertinoDialogAction(
+                                                            onPressed: () {
+                                                              FirebaseFirestore.instance
+                                                                  .collection("typeOfWashes")
+                                                                  .doc(snapshot
+                                                                  .data!.docs[index].id)
+                                                                  .delete();
+                                                              Navigator.pop(context);
+                                                            },
+                                                            child: const Text(
+                                                                "Yes"),
+                                                          ),
+                                                        ],
+                                                      );
+                                                    });
+
                                               },
                                               icon: Icon(
                                                 CupertinoIcons.delete_simple,
@@ -133,8 +163,6 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                                               0.02,
                                         ),
                                         child: Row(
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.center,
                                             children: [
                                               CachedNetworkImage(
                                                   fit: BoxFit.cover,
@@ -149,12 +177,12 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                                                                     context)
                                                                 .size
                                                                 .width *
-                                                            0.08,
+                                                            0.1,
                                                         width: MediaQuery.of(
                                                                     context)
                                                                 .size
                                                                 .width *
-                                                            0.08,
+                                                            0.1,
                                                         decoration: BoxDecoration(
                                                             borderRadius:
                                                                 BorderRadius
@@ -169,7 +197,10 @@ class _TypesOfWashesState extends State<TypesOfWashes> {
                                                   progressIndicatorBuilder: (context,
                                                           url,
                                                           downloadProgress) =>
-                                                      const CircularProgressIndicator(),
+                                                      Padding(
+                                                        padding:  EdgeInsets.only(left:MediaQuery.of(context).size.width*0.04, right:MediaQuery.of(context).size.width*0.04),
+                                                        child: const CircularProgressIndicator(),
+                                                      ),
                                                   errorWidget: (context, url,
                                                           error) =>
                                                       Padding(
